refactor design smells
