+++ src/main/java/org/apache/commons/collections4/list/CursorableLinkedList.java
@@ -516,18 +516,6 @@ public class CursorableLinkedList<E> extends AbstractLinkedList<E> implements Se
     }
 
     /**
-     * Deserializes the data held in this object to the stream specified.
-     *
-     * @param in  the input stream
-     * @throws IOException if an error occurs while reading from the stream
-     * @throws ClassNotFoundException if an object read from the stream can not be loaded
-     */
-    private void readObject(final ObjectInputStream in) throws IOException, ClassNotFoundException {
-        in.defaultReadObject();
-        doReadObject(in);
-    }
-
-    /**
      * Registers a cursor to be notified of changes to this list.
      *
      * @param cursor  the cursor to register
@@ -602,15 +590,5 @@ public class CursorableLinkedList<E> extends AbstractLinkedList<E> implements Se
         broadcastNodeChanged(node);
     }
 
-    /**
-     * Serializes the data held in this object to the stream specified.
-     *
-     * @param out  the output stream
-     * @throws IOException if an error occurs while writing to the stream
-     */
-    private void writeObject(final ObjectOutputStream out) throws IOException {
-        out.defaultWriteObject();
-        doWriteObject(out);
-    }
 
 }
