+++ src/main/java/org/apache/commons/collections4/list/NodeCachingLinkedList.java
@@ -170,18 +170,6 @@ public class NodeCachingLinkedList<E> extends AbstractLinkedList<E> implements S
     }
 
     /**
-     * Deserializes the data held in this object to the stream specified.
-     *
-     * @param in  the input stream
-     * @throws IOException if an error occurs while reading from the stream
-     * @throws ClassNotFoundException if an object read from the stream can not be loaded
-     */
-    private void readObject(final ObjectInputStream in) throws IOException, ClassNotFoundException {
-        in.defaultReadObject();
-        doReadObject(in);
-    }
-
-    /**
      * Removes all the nodes from the list, storing as many as required in the
      * cache for reuse.
      */
@@ -233,15 +221,5 @@ public class NodeCachingLinkedList<E> extends AbstractLinkedList<E> implements S
         }
     }
 
-    /**
-     * Serializes the data held in this object to the stream specified.
-     *
-     * @param out  the output stream
-     * @throws IOException if an error occurs while writing to the stream
-     */
-    private void writeObject(final ObjectOutputStream out) throws IOException {
-        out.defaultWriteObject();
-        doWriteObject(out);
-    }
 
 }
